requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
PyYAML==3.13
requests==2.19.1
urllib3==1.23
